feat!: separate cli and sharex script
